[simpleid] Use real random source to generate salt
Using a time stamp to derive randomness is critical as it allows informed
attackers to bruteforce the random information. While the implications in the
Uberspace setting might be manageable, it is still bad habit and should be
avoided at least for the sake of example.

This commit derives the salt from /dev/urandom, which is a cryptographically
sound source of randomness available on all Linux systems.