mailman: run qrunner as cronjob, not as a service
qrunner starts each of the 6 internal runners in an endless loop, eating up a full CPU core. This has been known since 2004, but apparently there is no real fix. The official setup guides use mailmalctl, which starts the runners independently of each other in a fire-and-forget fashion, sidestepping this bug.

https://mail.python.org/pipermail/mailman-users/2004-September/039401.html