Added External URL Support section
Changing the internal uber.space URL to an external URL with ghost command line as stated in the Ghost manual does not completely work as the users do not have root privileges to run sudo, the config files have to manually be updated as there is a change in the host IP address and the process has to be restarted with ``supervisorctl`` instead of the ghost-cli. Therefore I added this section to clarify such a change of the URLs.