Added crucial configuration step
Added a step how to modify the default .htaccess file to make Lychee work in the Uberspace environment. In addition, fixed two minor issues with the name of the composer executable and clarified how to use with a custom sub domain. Changed PHP version to 7.4 (default on Uberspace 7 right now).