Add Troubleshooting section for PHP Fatal error when updating to Nextcloud 21.0.1
Updating to Nextcloud 21.0.1 might fail returning a PHP Fatal error.
A single option in the php.ini file resolves the issue and allows
successfull update to Nextcloud 21.0.1.

This commit adds a section to the troubleshooting chapter in the Nextcloud Guide.
The section specifies the steps necessary to create the php.ini file,
set the needed apcu option and restart PHP to read the new configuration.

Resolves: #989