Fix(NOISSUE): Remove mentioning `nginx` from flask guide
The flask guide mentions "Configure nginx". Since the user doesn't
interact with nginx directly (I'm not even sure if Uberspace uses nginx, IIRC it's Apache), this was confusing to me.

I propose a change to "Uberspace web backend", as this phrasing is used in the Gitea documentation (which I found super helpful).