Update guide_linkstack.rst
Some minor but vital changes here: The database port needs to be localhost according to the Uberspace MySQL page, hence, the prior port number returned an error. LinkStack supports PHP 8.3, which is now included in this manual. Maybe more importantly, I have included to edit .htaccess in the likely use case of access though a subdomain. Lastly, the ambiguous registration section now recommends to set a secure admin registration right away, where it was formerly recommended to do so in a (unnecessary) second step.