Update guide_prosody.rst
Clarified how to handle the bug described in 
https://github.com/Uberspace/lab/issues/1234#issuecomment-2571435772

Removed coturn from Voice/Video Calls section as it doesn't work since ages and the coturn guide is removed from uberlab. 
Replaced it with eturnal.

Added me and don-philipe as authors because don-philipe gave an easy solution to handling the bug and me, because I tested, wrote down and added the eturnal thing.